+++ iati_account_web/data/models.py
@@ -3,8 +3,7 @@ from __future__ import annotations
 from datetime import datetime, timezone
 
 from django.db import models
-from iati_account_web.exceptions import RegisterYourDataResponseParsingIssue
-from iati_account_web.settings import (
+from iati_account_web.constants import (
     COUNTRY_LIST,
     LICENCE_LIST,
     ORGANISATION_TYPE_LIST,
@@ -13,6 +12,7 @@ from iati_account_web.settings import (
     USER_ROLE_LIST,
     VISIBILITY_LIST,
 )
+from iati_account_web.exceptions import RegisterYourDataResponseParsingIssue
 
 
 class UserAndRole(models.Model):