+++ iati_account_web/views.py
@@ -1,6 +1,7 @@
 import logging
 
 import libsuitecrm
+from django.conf import settings
 from django.http import HttpRequest, HttpResponse, HttpResponseRedirect
 from django.shortcuts import redirect
 from django.template import loader
@@ -11,7 +12,6 @@ from iati_account_web.metrics import (
     PROM_USER_PROVISIONING_COUNTER,
     PROM_USER_PROVISIONING_GAUGE,
 )
-from iati_account_web.settings import env
 from libsuitecrm import SuiteCRM
 
 app_logger = logging.getLogger("iati_account")
@@ -147,7 +147,7 @@ def __provision_add_roles(request: HttpRequest) -> bool:
 
     audit_logger.debug(f"Provisioning: Adding iati_register_your_data role to user {request.user.log_label}")
     try:
-        request.user.add_role_to_user_in_identity_service(env("IDENTITY_SERVICE_ROLE_ID_IATI_REGISTER_YOUR_DATA"))
+        request.user.add_role_to_user_in_identity_service(settings.IDENTITY_SERVICE_ROLE_ID_IATI_REGISTER_YOUR_DATA)
 
     except Exception as exc:
         audit_logger.error(f"Provisioning: Failed to add role for user {request.user.log_label} with exception {exc}")
@@ -174,7 +174,9 @@ def __provision_create_person_in_crm(request: HttpRequest) -> bool:
     audit_logger.debug(f"Provisioning: trying to create person in the CRM for user {request.user.log_label}")
     try:
         app_logger.debug("Provisioning: connecting to SuiteCRM")
-        crm = SuiteCRM(env("CRM_BASE_URL"), client_id=env("CRM_CLIENT_ID"), client_secret=env("CRM_CLIENT_SECRET"))
+        crm = SuiteCRM(
+            settings.CRM_BASE_URL, client_id=settings.CRM_CLIENT_ID, client_secret=settings.CRM_CLIENT_SECRET
+        )
         crm.fetch_access_token()
 
         app_logger.debug("Provisioning: creating Person record")