+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.4.18 - 2026-05-19
+
+### Changes
+
+- Refactor settings to use the best practice of importing settings from django.conf.
+- Refactor settings to move constant data (e.g., codelists, timezone lists) from settings.
+- Change database connection to configure from a single connection string URL, simplifying the configuration of the application.
+- Removed environment variables for the paths to codelist and licence JSON files, simplifying the configuration of the application.
+
+### Fixed
+
+- Small bug where "Superadmin" is listed as a possible user role in the list of users for an organisation.
+
+### Added
+
+- Unit test of some of the OIDC code.
+
 ## 0.4.17 - 2026-05-12
 
 ### Changes